Checkstyle fixes and refactor Name into its own package for tests
