+++ src/main/java/seedu/partyplanet/logic/commands/EditFieldCommand.java
@@ -13,10 +13,10 @@ import seedu.partyplanet.commons.core.index.Index;
 import seedu.partyplanet.commons.util.CollectionUtil;
 import seedu.partyplanet.logic.commands.exceptions.CommandException;
 import seedu.partyplanet.model.Model;
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Birthday;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Person;
 import seedu.partyplanet.model.person.Phone;
 import seedu.partyplanet.model.remark.Remark;
