+++ src/main/java/seedu/partyplanet/logic/parser/AddCommandParser.java
@@ -14,10 +14,10 @@ import java.util.stream.Stream;
 
 import seedu.partyplanet.logic.commands.AddCommand;
 import seedu.partyplanet.logic.parser.exceptions.ParseException;
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Birthday;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Person;
 import seedu.partyplanet.model.person.Phone;
 import seedu.partyplanet.model.remark.Remark;
