+++ src/main/java/seedu/partyplanet/logic/parser/ParserUtil.java
@@ -14,10 +14,10 @@ import seedu.partyplanet.commons.util.StringUtil;
 import seedu.partyplanet.logic.parser.exceptions.ParseException;
 import seedu.partyplanet.model.date.Date;
 import seedu.partyplanet.model.event.EventDate;
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Birthday;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Phone;
 import seedu.partyplanet.model.remark.Remark;
 import seedu.partyplanet.model.tag.Tag;
