+++ src/test/java/seedu/partyplanet/testutil/PersonBuilder.java
@@ -3,10 +3,10 @@ package seedu.partyplanet.testutil;
 import java.util.HashSet;
 import java.util.Set;
 
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Birthday;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Person;
 import seedu.partyplanet.model.person.Phone;
 import seedu.partyplanet.model.remark.Remark;
